hotfix: 랭킹 응답 수정

+++ server/userttubeot/src/main/java/com/user/userttubeot/user/domain/dto/UserRankDto.java
@@ -12,5 +12,5 @@ public class UserRankDto {
     private Integer userId;
     private String username;
     private Integer score;
-    private Long ttubeotId;
+    private Integer ttubeotId;
 }
